requirements.txt
pytest
mypy
pycodestyle
pyyaml
pandas